requirements.txt
exa-py>=1.0.0
pandas>=2.0.0
streamlit>=1.34.0
xlsxwriter>=3.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
